alterei o id de long para int

+++ trunk/fonte/ieadam/src/main/java/br/com/ieadam/repositorio/PerfilRepositorioSql.java
@@ -7,6 +7,6 @@ import br.com.ieadam.dominio.Perfil;
 public interface PerfilRepositorioSql {
 
 	public List<Perfil> listaPerfilPorSistemaPorUsuario(int sistemaId,
-			Long usuarioId);
+			int usuarioId);
 
 }
